+++ lib/shared/link_preview_card.dart
@@ -7,13 +7,14 @@ import 'package:url_launcher/url_launcher.dart';
 import 'package:thunder/shared/image_preview.dart';
 
 class LinkPreviewCard extends StatelessWidget {
-  const LinkPreviewCard({super.key, this.originURL, this.mediaURL, this.mediaHeight, this.mediaWidth, this.showLinkPreviews = true});
+  const LinkPreviewCard({super.key, this.originURL, this.mediaURL, this.mediaHeight, this.mediaWidth, this.showLinkPreviews = true, this.showFullHeightImages = false});
 
   final String? originURL;
   final String? mediaURL;
   final double? mediaHeight;
   final double? mediaWidth;
   final bool showLinkPreviews;
+  final bool showFullHeightImages;
 
   Future<void> _launchURL(url) async {
     Uri uri = Uri.parse(url);
@@ -25,7 +26,7 @@ class LinkPreviewCard extends StatelessWidget {
     final theme = Theme.of(context);
     final useDarkTheme = context.read<ThemeBloc>().state.useDarkTheme;
 
-    if (mediaURL != null && mediaHeight != null && mediaWidth != null) {
+    if (mediaURL != null) {
       return Padding(
         padding: const EdgeInsets.only(top: 4.0, bottom: 8.0),
         child: InkWell(
@@ -36,7 +37,13 @@ class LinkPreviewCard extends StatelessWidget {
               alignment: Alignment.bottomRight,
               fit: StackFit.passthrough,
               children: [
-                if (showLinkPreviews) ImagePreview(url: mediaURL!, height: mediaHeight, width: mediaWidth, isExpandable: false),
+                if (showLinkPreviews)
+                  ImagePreview(
+                    url: mediaURL!,
+                    height: showFullHeightImages ? mediaHeight : null,
+                    width: mediaWidth ?? MediaQuery.of(context).size.width - 24,
+                    isExpandable: false,
+                  ),
                 Container(
                   color: useDarkTheme ? Colors.grey.shade900 : Colors.grey.shade700,
                   padding: const EdgeInsets.symmetric(horizontal: 8.0, vertical: 12.0),